+++ src/test/java/UserCreateWrongCredsTest.java
@@ -1,5 +1,4 @@
 import burgerapi.UserClient;
-import dto.UserCreate;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import model.User;
@@ -12,31 +11,31 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 public class UserCreateWrongCredsTest {
-    User user;
     UserClient userClient;
     private String accessToken;
-    private String refreshToken;
 
     @Before
     public void setUp(){
         userClient = new UserClient();
-        user = UserCreate.getRandomUser();
-        userClient.create(user);
+        //user = User.getRandomUser();
+        //userClient.create(user);
 
-        ValidatableResponse loginResponse = userClient.login(user);
+       /* ValidatableResponse loginResponse = userClient.login(user);
         accessToken = loginResponse.log().all().extract().path("accessToken"); //.toString()
-        refreshToken = loginResponse.extract().path("refreshToken"); //.toString()
+        refreshToken = loginResponse.extract().path("refreshToken"); //.toString()*/
     }
 
     @After
     public void tearDown(){
-        userClient.delete(accessToken);
+        if (accessToken != null) {
+            userClient.delete(accessToken);
+        }
     }
 
     @Test
     @DisplayName("Создание неуникального пользователя")
     public void createUserWithoutOneFieldTest() {
-        ValidatableResponse createResponse = userClient.create(user);
+        ValidatableResponse createResponse = userClient.create(User.getRandomUserWithoutEmail());
         int statusCode = createResponse.log().all().extract().statusCode();
         boolean responseOk = createResponse.log().all().extract().path("success");
         assertThat("409", statusCode, equalTo(SC_FORBIDDEN));
