+++ src/test/java/UserUpdateTest.java
@@ -1,10 +1,7 @@
 import burgerapi.UserClient;
-import dto.UserCreate;
-import dto.UserLogin;
 import io.restassured.response.ValidatableResponse;
 import model.User;
 import io.qameta.allure.junit4.DisplayName;
-import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -18,27 +15,26 @@ import static org.hamcrest.MatcherAssert.assertThat;
 public class UserUpdateTest {
     User user;
 
-    UserLogin userLogin;
     UserClient userClient;
-    UserCreate userCreate;
 
     private String accessToken;
-    private String refreshToken;
 
     @Before
     public void setUp(){
         userClient = new UserClient();
-        user = UserCreate.getRandomUser();
+        user = User.getRandomUser();
         userClient.create(user);
 
         ValidatableResponse loginResponse = userClient.login(user);
         accessToken = loginResponse.log().all().extract().path("accessToken"); //.toString()
-        refreshToken = loginResponse.extract().path("refreshToken"); //.toString()
+
     }
 
     @After
     public void tearDown(){
-        userClient.delete(accessToken);
+        if (accessToken != null) {
+            userClient.delete(accessToken);
+        }
     }
 
     @Test
